functions/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv
pydantic
firebase-functions
firebase-admin
a2wsgi
